Software/Python/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
 
[project]
name = "educational_robot"
version = "0.1.0"
description = "STS3215 servo robot arm driver and API"
requires-python = ">=3.10"
dependencies = [
    "pyserial",
]
 
[tool.setuptools.packages.find]
where = ["."]
include = ["robot_arm*", "robot_api*"]
 